# core/gen/api.gen.go
//go:build go1.22

// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package gen

import (
	"fmt"
	"net/http"

	"github.com/oapi-codegen/runtime"
)

// Defines values for EdgeConfidence.
const (
	Exact    EdgeConfidence = "exact"
	Probable EdgeConfidence = "probable"
)

// Valid indicates whether the value is a known member of the EdgeConfidence enum.
func (e EdgeConfidence) Valid() bool {
	switch e {
	case Exact:
		return true
	case Probable:
		return true
	default:
		return false
	}
}

// Defines values for EdgeKind.
const (
	Defines    EdgeKind = "defines"
	References EdgeKind = "references"
)

// Valid indicates whether the value is a known member of the EdgeKind enum.
func (e EdgeKind) Valid() bool {
	switch e {
	case Defines:
		return true
	case References:
		return true
	default:
		return false
	}
}

// Defines values for NodeKind.
const (
	File   NodeKind = "file"
	Symbol NodeKind = "symbol"
)

// Valid indicates whether the value is a known member of the NodeKind enum.
func (e NodeKind) Valid() bool {
	switch e {
	case File:
		return true
	case Symbol:
		return true
	default:
		return false
	}
}

// Edge defines model for Edge.
type Edge struct {
	Confidence EdgeConfidence `json:"confidence"`

	// From Source node id
	From string   `json:"from"`
	Id   string   `json:"id"`
	Kind EdgeKind `json:"kind"`

	// To Target node id
	To string `json:"to"`
}

// EdgeConfidence defines model for Edge.Confidence.
type EdgeConfidence string

// EdgeKind defines model for Edge.Kind.
type EdgeKind string

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
}

// FileContent defines model for FileContent.
type FileContent struct {
	Content string `json:"content"`
	Path    string `json:"path"`
}

// Graph defines model for Graph.
type Graph struct {
	Edges []Edge `json:"edges"`
	Nodes []Node `json:"nodes"`
}

// Node defines model for Node.
type Node struct {
	Id    string   `json:"id"`
	Kind  NodeKind `json:"kind"`
	Label string   `json:"label"`

	// Path Absolute file path (file node) or containing file (symbol node)
	Path  *string `json:"path,omitempty"`
	Range *Range  `json:"range,omitempty"`
}

// NodeKind defines model for Node.Kind.
type NodeKind string

// Position defines model for Position.
type Position struct {
	// Character 0-based UTF-16 character offset (LSP convention)
	Character int `json:"character"`

	// Line 0-based line number (LSP convention)
	Line int `json:"line"`
}

// Range defines model for Range.
type Range struct {
	End   Position `json:"end"`
	Start Position `json:"start"`
}

// GetFileParams defines parameters for GetFile.
type GetFileParams struct {
	// Path Absolute path to the file within the analyzed root
	Path string `form:"path" json:"path"`
}

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get file content
	// (GET /file)
	GetFile(w http.ResponseWriter, r *http.Request, params GetFileParams)
	// Get dependency graph
	// (GET /graph)
	GetGraph(w http.ResponseWriter, r *http.Request)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetFile operation middleware
func (siw *ServerInterfaceWrapper) GetFile(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetFileParams

	// ------------- Required query parameter "path" -------------

	if paramValue := r.URL.Query().Get("path"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "path"})
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", true, true, "path", r.URL.Query(), &params.Path, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "path", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetFile(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetGraph operation middleware
func (siw *ServerInterfaceWrapper) GetGraph(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetGraph(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of http.ServeMux.
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	ServeHTTP(w http.ResponseWriter, r *http.Request)
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc("GET "+options.BaseURL+"/file", wrapper.GetFile)
	m.HandleFunc("GET "+options.BaseURL+"/graph", wrapper.GetGraph)

	return m
}